refactor: split post detail components and apply Suspense
- 게시글 상세 페이지의 컴포넌트 분리
- Suspense 적용
