+++ src/domains/post/detail/_hooks/usePostDetail.ts
@@ -1,4 +1,7 @@
-import { useQuery } from '@tanstack/react-query';
+'use client';
+
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { useParams } from 'next/navigation';
 
 import { BASE_URL } from '@lib/api/url';
 import { Post } from '@lib/posts/postFactory';
@@ -18,11 +21,14 @@ async function fetchPostDetail({ category, slug }: Props): Promise<Post> {
   return res.json();
 }
 
-export default function usePostDetail({ category, slug }: Props) {
-  return useQuery({
+export default function usePostDetail() {
+  const params = useParams();
+  const category = params.category as string;
+  const slug = params.slug as string;
+
+  return useSuspenseQuery({
     queryKey: ['post', { category, slug }],
     queryFn: () => fetchPostDetail({ category, slug }),
     staleTime: 1000 * 60 * 5,
-    enabled: Boolean(slug) && Boolean(category),
   });
 }
